+++ src/core/entry/compiler/target.rs
@@ -3,20 +3,18 @@ use std::{
     path::{Path, PathBuf},
     str::FromStr,
 };
-
-use crate::core::util::real_chain_env_toml;
 use clap::ValueEnum;
 use gen_utils::{
     common::{DepType, RustDependence, Source, ToToml},
     compiler::{CompilerImpl, UnderlayerConfImpl},
-    error::{Error, ParseError},
+    error::Error,
 };
 use makepad_gen_plugin::compiler::{
     Compiler as MakepadCompiler, Config as MakepadConfig,
     CONF_FORMAT_SUGGESTION as MAKEPAD_CONF_FORMAT_SUGGESTION,
 };
 use toml_edit::{DocumentMut, Formatted, InlineTable, Item, Value};
-
+use crate::core::entry::ChainEnvToml;
 use super::GenUIConf;
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, ValueEnum)]
@@ -105,17 +103,11 @@ impl TryFrom<(DocumentMut, Underlayer)> for CompileUnderlayer {
                 || Err(Error::from(MAKEPAD_CONF_FORMAT_SUGGESTION)),
                 |table| {
                     // before to makepad config, add gen_components dependence into [makepad.dependencies]
-                    let gen_dep_path = PathBuf::from_str(
-                        real_chain_env_toml()?["dependencies"]["gen_components"]
-                            .as_str()
-                            .expect("gen-components path not found"),
-                    )
-                    .map_err(|e| {
-                        Error::Parse(ParseError::new(
-                            e.to_string().as_str(),
-                            gen_utils::error::ParseType::Toml,
-                        ))
-                    })?;
+                    let chain_env_toml: ChainEnvToml = ChainEnvToml::path()?.try_into()?;
+                    let gen_dep_path = chain_env_toml.gen_components_path().map_or_else(
+                        || Err(Error::from("can not find [dependencies.gen_components] in env.toml, maybe config broken, use `ract init` to fix it")),
+                        |f| Ok(f),
+                    )?;
 
                     table
                         .as_table_mut()
@@ -152,18 +144,12 @@ impl TryFrom<(DocumentMut, Underlayer)> for CompileUnderlayer {
 impl CompileUnderlayer {
     /// default makepad toml
     fn makepad(root: PathBuf) -> Result<Self, Error> {
-        let chain_env_toml = real_chain_env_toml()?;
-        let makepad_dep_path = PathBuf::from_str(
-            chain_env_toml["dependencies"]["makepad-widgets"]
-                .as_str()
-                .expect("makepad-widgets path not found"),
-        )
-        .map_err(|e| {
-            Error::Parse(ParseError::new(
-                e.to_string().as_str(),
-                gen_utils::error::ParseType::Toml,
-            ))
-        })?;
+        let chain_env_toml: ChainEnvToml = ChainEnvToml::path()?.try_into()?;
+        let makepad_dep_path = chain_env_toml.makepad_widgets_path().map_or_else(
+            || Err(Error::from("can not find [dependencies.makepad-widgets] in env.toml, maybe config broken, use `ract init` to fix it")),
+            |f| Ok(f),
+        )?;
+
         let mut makepad_dep = RustDependence::new("makepad-widgets");
         makepad_dep.set_ty(DepType::Local(makepad_dep_path.join("widgets")));
         // new makepad config and add dependence